feat: refactor color management and streamline color services
This commit refactors the structures and services related to color management. The "DynamicColorKey" and default color definitions have been moved to a separate model file, streamlining the ColorManagerService. A significant part of the color handling logic is also cleaned up, reducing repetition and improving maintainability of the code. Unnecessary code blocks were removed as well.